dash_data_retention/requirements.txt
pyrebase4==4.5.0
pyparsing<3,>=2.0.2
